requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
SQLAlchemy~=1.3.0
requests==2.27.1
flask-redis==0.4.0
waitress==1.4.3
pyjwt==1.7.1
redis==4.3.4
